code review
Closes gh-6763
