+++ micrometer-test/src/main/java/io/micrometer/test/assertions/TimerAssert.java
@@ -13,14 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micrometer.common.util.assertions;
+package io.micrometer.test.assertions;
 
 import java.time.Duration;
 import java.util.concurrent.TimeUnit;
 import java.util.function.Function;
 
 import org.assertj.core.api.AbstractAssert;
-import org.assertj.core.api.DoubleAssert;
 import org.assertj.core.api.DurationAssert;
 import org.assertj.core.api.IntegerAssert;
 
@@ -34,6 +33,7 @@ import io.micrometer.core.instrument.Timer;
  * use {@link io.micrometer.core.tck.MeterRegistryAssert#timer(String, Tag...)}.
  *
  * @author Emanuel Trandafir
+ * @since 1.17.0
  */
 public class TimerAssert extends AbstractAssert<TimerAssert, Timer> {
