+++ micrometer-test/src/test/java/io/micrometer/test/assertions/TimerAssertTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micrometer.common.util.assertions;
+package io.micrometer.test.assertions;
 
 import io.micrometer.core.instrument.Tag;
 import io.micrometer.core.instrument.Timer;
